requirements.txt
# ─────────────────────────────────────────
# FRAMEWORK PRINCIPAL
# ─────────────────────────────────────────
Flask==3.0.3

# ─────────────────────────────────────────
# BASE DE DONNÉES
# ─────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
Flask-Migrate==4.0.7

# ─────────────────────────────────────────
# AUTHENTIFICATION & SÉCURITÉ
# ─────────────────────────────────────────
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3
email-validator==2.2.0

# ─────────────────────────────────────────
# EMAILS
# ─────────────────────────────────────────
Flask-Mail==0.10.0

# ─────────────────────────────────────────
# UPLOAD FICHIERS
# ─────────────────────────────────────────
Pillow==10.4.0

# ─────────────────────────────────────────
# VARIABLES D'ENVIRONNEMENT
# ─────────────────────────────────────────
python-dotenv==1.0.1

# ─────────────────────────────────────────
# TESTS
# ─────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0

# ─────────────────────────────────────────
# QUALITÉ DE CODE
# ─────────────────────────────────────────
flake8==7.1.0
black==24.4.2